requirements.txt
# === Core ===
numpy
pandas
scikit-learn
joblib
python-dotenv

# === Deep Learning ===
tensorflow
keras
scikeras

# === API - FastAPI ===
fastapi
uvicorn

# === Interface - Streamlit ===
streamlit

# === Visualisation ===
matplotlib
seaborn

# === Google Cloud Platform ===
google-cloud-storage
google-cloud-bigquery
google-auth
google-auth-oauthlib
google-api-core

# === Dev / Test / Lint ===
ipykernel
pytest
black
isort
pylint
tqdm
